requirements.txt
Django==2.2.4
django-extensions==2.2.1
django-shell-plus==1.1.7
djangorestframework==3.10.2
pylint==2.3.1
